requirements.txt
alpaca-py==0.42.0
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
msgpack==1.1.1
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
plotly==6.2.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sseclient-py==1.8.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
